Default values for rectangle detection in iOS 11.x and later
The default value for VNDetectRectanglesRequest::minimumAspectRatio is 0.5.
What this means is that any document less than a 1:2 ratio will fail.

Suggest we use a default value of at least 0.1
I see there is a feature request for passing various options.
https://github.com/WeTransfer/WeScan/issues/81#issue-380241097

Perhaps passing a minimumAspectRatio would suit the needs of more applications.