Adjust `HomeViewController` for iOS 13.
The modal wasn't working great when not presented in full screen because of the drag down to dismiss gesture.
We're now presenting in full screen to fix this issue, and also because it makes more sense for camera screens.
We also fixed a problem with dark mode where the navigation bar would be black with black buttons.